Move dll import codes to win32 helper
